Worked on script to transfer store collections
Initially, I created a script to store and manipulate data in the collections. I stored this data in a JSON file. For the sake of easier data manipulation and avoiding spaghetti code. This data will further be stored in an SQLite database,
TODO// copy functionality within jsonStore.py file into sqliteStore.py.