Replaced Existing JSON storage functionality
tinkered with SQLite commands. Inserting a new collection or app path will no longer duplicate it within the collections table or apps table, respectively.

Also, added functionality to return all apps alongside their collections from within the database.  This is to be implemented with a running function which continually updates the canvas to show the active collections and their corresponding apps.